Fixed FilterObject
