+++ SEDaily-IOS/CustomTabViewController.swift
@@ -92,10 +92,8 @@ class CustomTabViewController: UITabBarController, UITabBarControllerDelegate {
     func setupTabs() {
         let layout = UICollectionViewLayout()
         let vc1 = PodcastPageViewController()
-        let vc2 = JustForYouCollectionViewController(collectionViewLayout: layout)
-        let vc3 = TopCollectionViewController(collectionViewLayout: layout)
-//        let vc2 = GeneralCollectionViewController(collectionViewLayout: layout, type: API.Types.recommended)
-//        let vc3 = GeneralCollectionViewController(collectionViewLayout: layout, type: API.Types.top)
+        let vc2 = GeneralCollectionViewController(collectionViewLayout: layout, type: .recommended)
+        let vc3 = GeneralCollectionViewController(collectionViewLayout: layout, type: .top)
         
         let icon1 = UITabBarItem(title: L10n.tabBarTitleLatest, image: #imageLiteral(resourceName: "mic_stand"), selectedImage: #imageLiteral(resourceName: "mic_stand_selected"))
         let icon2 = UITabBarItem(title: L10n.tabBarJustForYou, image: #imageLiteral(resourceName: "activity_feed"), selectedImage: #imageLiteral(resourceName: "activity_feed_selected"))
